﻿ENH: Removing the hacks introduced in r2081 and r5516 and applying a better solution to give the proper orientation when making sagital or coronal reconstructions.
Instead of the hack of r2081 and r5516, the approach is to use the setImageOrientation() method after the view has been reseted, so no patient position checks are needed, so the solution is more general for any case, giving
always the standard coronal or sagital orientation as a result.
This clean solution has been possible thanks to the changes introduced in r8898.