BUG: Modify the generation of the texture in VtkTextActorWithBackground to fix distortion problems in some cases.
With these changes we ensure that the output of m_imageBlend, which is used as the source of the texture, has always dimensions that are powers of 2. This is achieved by creating a background image with the needed texture size, and reducing the original ImageData to just fit the text and translating it to account for the margin.