BUG: Check that the expected attributes in a VOI LUT sequence item really exist.
Until now it was assumed that LUT Descriptor, LUT Explanation and LUT Data were all present, but LUT Explanation is of type 3, i.e. optional, so it may not be present. Also it is now checked that the mandatory attributes exist, and the item is skipped if they aren't present.