ENH: Rework the Q2DViewerWidget to ensure a proper visual clean up of the Viewers Layout when viewers are deleted.
Q2DViewerWidget is now a QStackedWidget with the first page being the equivalent to the previous Q2DViewerWidget and the second page being a black widget. Before deleting Q2DViewerWidget the second page is set as current and a repaint is forced, this way ensuring that there isn't any leftover of the deleted QVTKWidget in the viewer and a solid black color is painted instead.