BUG: Fix ordering of images in some cases.
Since Qt5 all the QVector3D methods accept and return floats instead of qreals (which are doubles in desktop platforms) like in Qt4. In this specific case, the change from Qt4 to Qt5 implied that the intermediate operations in angleInRadians() were until now made with float precision instead of double precision like before, and in some cases the same angle was obtained in cases where different angles were obtained before. This affected image ordering producing an image order different from the expected in test T00117. By casting all the values to double we revert back to the old intermediate precision and obtain the same image order than with Qt4.