ENH: Implement thick slab using the VTK reslice mapper.
Conceptual changes:

- Thick slab defined in mm instead of number of slices.
- New projection mode from VTK: sum.

The new class SliceOrientedVolumePixelData has been introduced to abstract from coordinate changes in every place where it has been possible. This is necessary because the thick slab pixel data is now obtained from the mapper and is very different from the original in terms of axes orientations and origin.