ENH: Slicing tool implementation in floating point numbers.
Slicing tools and it's child classes have been refactored to implement the
more generalistic approach of floating point numbers instead of integers when
setting the location for an slice.

If the current implementation does not support or its nature does not
allow it (like volumes) then rounding is used.

setLocation and incrementLocation methods return the amount of "unused"
scrolling distance in float, so when a number is rounded, arrives at a limit,
etc... then this accounts as "unused" distance.

The implementation takes in account the nature of the floating points
operations: usage of epsillons, not using == statements in floats,
rounding, "magnet effect" when arriving at the limits at an epsillon
distance, etc...