ENH: Improve the crash button handling.
In the previous implementation the ok button ceased to work after the crash
button was shown, and several crash buttons could be created.

Now there's only one crash button that is hidden by default and it shows on the
left of the other buttons. The ok button has been replaced by a close button and
it's the one with the focus by default.

Additionally, the WA_DeleteOnClose attribute has been set.