BUG: Rework a unit test to reduce random crashes in macOS.
Sometimes autotests crashes at the end in macOS. Apparently it has
something to do with the destruction of the QObject hierarchy, so we
have devised a method to destruct the VolumePixelData in
test_VolumePixelDataIterator without relying on it being destroyed
automatically when its parent is destroyed.

This should should reduce the frequency of the crashes, but they still
can happen.