ENH: Scary warning for cryptographically invalid downloads #2768
Now, the user is suggested to delete the downloaded file when an invalid download is detected.

The motivation to do so is to avoid the potentially maliscious code to become compiled in case the scripts are invoked one by one.

Moreover, an scary warning explaining the situation is provided to the user. A situation like that is a symptom of something bad happening.