COMP: “Fuck the system” for the new Qt mandatory login installation.
Albeit the downloaded installer binary did not change, it seems it always called home to check if it had to nag you.
Those Microsoftesque practices (search how to create a local account on windows 10) will only help to increase the level of distrust towards Qt.

Starviewer is a GPLv3 project, fully complaint with Qt's licensing terms.

Having to use a Qt account every time we recompile from scratch is a hassle, especially when one does it from “clean” environments.
This commit tricks the installer to not see the internet by setting the environment variable https_proxy to an unreachable proxy server (0.0.0.0).