BUG: Fix crash at autotests end.
autotests was crashing at its end because of a combination of factors
involving static storage.

1. On one side there is the itk::TimeStamp class (used in itk::Object)
with a static member named m_GlobalTimeStamp. VolumePixelData contains
two ITK objects, thus it ultimately references that static object.

2. On the other side our DECLARE_TEST macro creates a static instance of
the test class given as argument. We had several Volume and
VolumePixelData instances as children of some of the test classes that
were deleted only when their parent (a static instance) is deleted.

3. Destruction order of static objects is the reverse of construction
order, which is too hard to control. In this case m_GlobalTimeStamp was
deleted before the test class instances, thus, when the VolumePixelData
instances were deleted they tried to access a deleted object (a null
smart pointer to be more precise) and a segmentation fault was ensued.

The solution has been to ensure destruction of all Volume and
VolumePixelData instances either with a manual delete or using smart
pointers, depending on each case. Additionally, VolumePixelData is no
longer a QObject subclass (it didn't profit from it) and some subtle
memory leaks have been fixed.