COMP: Avoid top level const parameters in function declarations. #2844
Detected with clang-tidy. In a parameter passed by value it doesn't
matter if it is declared as const or not in the function declaration,
the function signature is the same; it only matters in the function
definition.

In some cases the parameter has been changed to a const reference and in
others just the const qualifier for the parameter in the function
declaration has been removed.