ENH: Implement propagation of VOI LUT from any input in a viewer. #2151
A new AnyInputSyncCriterion has been created that is satifisfied when
there's any coincidence between the inputs of the two given viewers.
VoiLutSignalToSyncActionMapper has been modified to iterate over all the
inputs of the source viewer and VoiLutSyncAction has to check that the
target viewer actually contains the source volume (it can happen that
the source viewer has two inputs and the target just one and
VoiLutSyncAction will be called for both source inputs). Finally, the
WindowLevelTool has to set the explanation with the custom preset name
so that it's classified correctly by VoiLutPresetsToolData (it wasn't
necessary before because when it modifies the main input the code
follows a more complex path that sets this explanation in another
place).