ENH: Automatic MSYS and MSVC installation and configuration.
The install_dependencies.sh has been renamed to configure_system.sh.

A very automated process in order to succesfully install and configure a Windows system development machine has been implemented.

MSVC build tools are automatically downloaded, an offline installer is created and then this offile installer is used to perform the installation.

MSYS packages are updated and installed using Pacman.

Also, the user is asked if an automatic enablement of long paths Windows 10 feature has to be performed. This is important as ITK compilation will fail.

User interaction is required in some parts due to technical limitations (the user is guided through all the quirks). Windows things…  ¯\_(ツ)_/¯