ENH: Rename variables and assume one type of SDK compilation.
The SDK could be compiled in different modes (debug, release…) in the same SDK. However, when the SDK is installed, only the last to build was usable by Starviewer. Now, only one type of compilation can be specified.

Some SDK variables have been renamed to better express their relationship with the SDK.