ENH: Support for SDK in debug mode on Windows.
Allow Starviewer to use a debug mode compiled SDK. For this reason, changes to fixdebug had to be done to apply it only when the SDK is in release mode.

When the SDK is compiled in debug mode, the linking has to be to the system MTd library instead of MT library, this is now automatically detected and applied.