ENH: Improved generation of prefix.sh and prefix.bat
A list of variables to be exported is stored in a list, and then the generate_prefixfiles.sh makes the corresponding exporting scripts and .txt files to ease the environment portability (copy and paste to Qt Creator).

The SDK library configuration scripts are interated in setup.sh to make the library-specific libraries always available.