ENH: Read local studies from RelatedStudiesManager without going through QueryScreen. #2764
This avoids the problem of a related study opening in another window for
the case of local studies (it still remains for the case of remote
studies). The problem was introduced when the logic to assign patients
to windows was adjusted during #2732.

This change also makes the path simpler for local related studies and
removes some responsibility from QueryScreen, which is good.

The method generatePatientVolumes, until now exclusive to
ExtensionHandler, has been moved to VolumeHelper to make it available in
different places.