ENH: Read color functions from resources just the first time. #2132
Profiling we have found that 38.6% of the time to apply the next hanging
protocol in sequence (with F12 shortcut) was spent in the
TransferFunctionModelFiller::add2DTransferfFunctions, which is called
once for every created viewer. Most of this time was spent reading the
functions from XML.

Thus we now read them only the first time and store them in a vector
from which we fill the given TransferFunctionModel without need to read
any file. This change reduces layout changing time in general, not only
for hanging protocols.