BUG: Parameter binding with position didn't always work.
The consequence was that the database revision was not updated.

We are sure it has not worked on Windows 10 with Qt 5.12.6 MSVC2017 64-bit but it has worked in Debian 10 stable with Qt 5.12.6 gcc from the website. Thus we don't know the specific circumstances in which it failed but it should have worked always according to the Qt documentation. Hopefully with the change it works in all cases.