ENH: Rework VolumeReaderJobFactory::read to avoid race conditions. #2680
With the previous design the read method created, enqueued and returned
a VolumeReaderJob and then the client (VolumeReaderManager) had to
connect to the job's signals. It could happen (though unlikely) that the
job finished and emitted all its signals before those connections where
made. That's the race condition we talk about.

With the new design the read method does not return anything but instead
requires a requester parameter. It creates the job, connects the job's
signals to this class' (VolumeReaderJobFactory) slots, takes note of the
requester and then enqueues the job. In the slots it emits a
corresponding signal for each requester of a volume. This way the client
(VolumeReaderManager) only has to connect to this class' signals before
reading and on its slots check that the requester matches.

It's the same approach recently used for PacsManager.