ENH: Alternative flip implementation. #2874
Instead of using Azimuth(180), which uses a rotation transformation to
alter the camera position, now we modify it directly. We compute a point
reflection through the focal point using a simple vector addition. This
is more numerically stable and allows to use the fast path (permute
execute) in thick slab calculation.