ENH: Implement a 2D polydata mapper with stippled lines support. #2879
It is a subclass of vtkOpenGLPolyDataMapper2D that implements stippled
lines with a GLSL shader.

The idea comes from a comment in VTK issue 15799.

For the moment this new mapper is only used in DrawerPolyline, which is
the only DrawerPrimitive that is currently used with stippled lines at
least once.