ENH: New UI for thick slab. #2896
The submenu widget has been transformed in a widget on the side that can
be shown or hidden. The button with its side widget have been moved to
the right end of the toolbar to avoid moving other buttons when the side
widget is toggled. In addition, the whole toolbar is in a scroll area so
it can fit in small screens.