ENH: Make the thick slab widget foldable only when the window is small. #2896
Now the thick slab widget can be foldable or not. When not foldable (the
default) the options widget is always visible and the arrow button not.
When foldable it works like in the previous commit. Q2DViewerExtension
decides when it must be foldable and when not according to the window
width: it is foldable only if scrollbars would appear otherwise.

This way for most people [citation needed] the options will be always
visible and thus more easily accessible.