BUG: Check if each object exists in the db to decide if insert or update. #2757
Until now it was always tried to do an insert and if it failed with a
constraint error it was assumed that it was because it already existed
and an update was tried instead. Now it is actually checked if the
object exists to decide whether to do an insert or an update. This is a
safer approach.

In addition, there was a case of shutters getting duplicated if their
parent image was updated due to an error when migrating from native
sqlite to QtQSL (they were inserted twice in case of update). This has
been fixed as well.