ENH: Implement #2962 storing a pointer to the AutomaticSynchronizationTool in Q2DViewer.
The tool is used to set the already synchronized slice when changing the
view plane. If there is no valid synchronization, the default slice
(first or middle) is used.

To allow for the reset tool to reset the slice, the pointer to the tool
is temporarily set to nullptr during the reset. In addition, the tool is
"disabled" to avoid synchronization in the acquisition plane even though
the viewer could be restored to another plane with the viewer command.

As a side effect, #2730 is also partially implemented: the last slice in
each plane is remembered if automatic synchronization is enabled. #2045
is also fixed by this change.

This solution introduces coupling between the viewer and the automatic
synchronization tool but it is the less dirty way to implement the
requested behaviour in short term. To implement a proper solution a very
deep overall redesign is required.