ENH: Corrections to allow building with CMake on Windows. #2828
Currently it works to build Starviewer in release mode and probably also in debug with debug libraries. To build Starviewer in debug with release libraries like we did until now, additional changes are required.

Also it's a bit cumbersome to build with MSYS, MSVC and jom, and more adaptation is needed to allow building with MinGW, but we are now closer to that goal.