ENH: Refactor setttings classes to ease using testing settings. #1975
The Settings class has been converted to the top level class to be used
just like always but its implementation has changed considerably. It no
longer inherits from SettingsInterface, but instead it contains a
pointer to an instance of SettingsInterface and delegates a big part of
its work to it. This pointer is initialized to either the value of a
similar static member, that allows injecting a TestingSettings, or in
normal conditions to a new instance of PersistentSettings.

This new class PersistentSettings is a subclass of SettingsInterface
that implements all the methods that use QSettings that were previously
defined directly in Settings.

Meanwhile, SettingsInterface has been extended with a few more methods
regarding lists that throw exceptions, just to avoid having to implement
them in TestingSettings for the moment. They will be implemented when
needed.

Overall, this redesign makes it easy to use testing settings without
having to modify any existing class. This scenario is demonstrated in
the updated test_VolumePixelDataReaderFactory.