BUG: Don't create a widget outside the main thread to query OpenGL version and extensions. #2920
It is unsafe and fails an assert with debug versions of Qt. This OpenGL
information is now queried in a more modern way and using a
QOffscreenSurface to avoid creating a widget. However on Windows Qt
still needs to create an invisible QWindow and thus can only be used
from the main thread. Thus SystemRequirementsTest will fail on Windows
but this will be fixed in the next commit.