BUG: Check that the system supports the required OpenGL version in a multiplatform safe way. #2920
The new approach requests a context with the minimum required version.
If it succeeds and the created context has an equal or greater version
the test is passed. Otherwise it's failed. This implementation does not
need a window nor an offscreen surface and fixes the problem left in the
previous commit.