ENH: Change separator for the default PACS list to query. #2866
Now it will be separated with backslashes instead of forward slashes.
Forward slashes were a problem for base URIs and can be used in AE
titles. The backslash is not allowed in AE titles and will be percent
encoded in URLs, thus it won't be a problem. The migration to the new
separator is performed on the first read.