BUG: Fix several problems when reading overlays.
Problem 1 (#2154): very long time to read overlays in multiframes.
Problem 2 (#2155): multiframe overlays ignore Image Frame Origin.
Problem 3 (#2986): overlay origin read as if it was x\y instead of
row\column.

Overlays are now read with DCMTK, allowing to read only one frame of the
file each time. This greatly reduces the reading time, although there's
still room for improvement. The other problems have also been fixed when
doing the new implementation.

However a new limitation has been introduced (#2987) due to a bug in the
currently used version of DCMTK (3.6.5). The limitation is that all
overlays are cropped to fit in the image confines. This will be fixed
when we upgrade DCMTK to a newer version (DCMTK 3.6.6 has already fixed
its bug).