BUG: DrawerPrimitive's destructor is now responsible for emitting the "dying" signal instead of delegating this responsibility to the subclasses' destructor (at the same time, this fixes a bug with DrawerPoint, whose destructor did not include the signal). #3000
When applying this change, another bug has appeared: temporary primitives are not removed from Drawer, such as the polylines of the Polygonal ROI and the eraser's rectangle. To solve this problem, a new attribute is created for DrawerPrimitive (a vtkProp) which is returned by DrawerPrimitive::getAsVtkProp. The DrawerPrimitive's subclasses fill the parent's attribute and call the parent's DrawerPrimitive::getAsVtkProp to return the vtkProp.