BUG: Assign the study received through ExternalStudyRequestManager to only the last active main window.
After the changes done to implement #2734 #2764 received studies were
assigned to all windows, thus failing test T00097. To solve it, we have
added a new ExternalStudyRequestMediator which is a singleton that
communicates with ExternalStudyRequestManager (also a singleton) and
then assigns the result to the last active Starviewer main window.

The study view or load operation still has to go through the window's
query screen, so this can still be further improved in a future
redesign.