ENH: New colours for DrawerPrimitives, reference lines and MPR to meet requirements.
All these colours have been chosen with the requirement of having a
noticeable brightness in each channel, meaning that each component is
distinct enough from black, for the reasons described in #3031.

In the case of DrawerPrimitives and reference lines we have chosen
colours that are similar to the previous ones.

For the MPR we have chosen a set of 3 colours that are different in each
channel, choosing 3 values and rotating them, so that whatever the
chosen channel they have 3 different gray levels.