BUG: Fix crash when cancelling studies downloaded with WADO-URI. #3061
There was access to m_pendingReplies at the same time from two threads,
one of them modifying it.

To avoid this problem, WADO-URI network replies are now aborted through
signal/slot connections. Since the replies live in the WADO thread and
the WadoUriRequest lives in the main thread, this is a queued
connection.

For all the other WADO requests abort is also called from the WADO
thread (with QMetaObject::invokeMethod) to avoid possible problems (it's
not documented as a thread-safe method).

There was another problem specific to WADO-URI: if the cancel request
was made before the actual request for each instance, then it was not
cancelled. Now the cancellation contemplates also cancellation during
the C-FIND (also with signal/slot connections), so that cancellation
works always. This has implied declaring the cancel() method in
StudyOperationResult and all its subclasses as a public slot.

Finally, a processEvents() in DirectoryUtilites has also been removed to
avoid another source of crashes during cancellation. This was harder to
track but it cause altered processing order of interthread signals and
eventually accessing destroyed objects.

We have also documented (in a comment in StudyOperationsService's
destructor) a possible crash when closing Starviewer while a C-FIND
operation is in progress. It is impossible to reproduce deterministicly,
but sometimes processing of a C-FIND query may continue after main() has
exited.

In addition to all this, some more methods that are run in the WADO
thread in WadoRequest subclasses have been documented as so. All methods
running in the WADO thread should now be properly marked.