ENH: Reading DICOM only when magic bytes match. #3088
Instead of reading any file as if it were a DICOM (which is costly), we only do
so if the DICOM magic bytes are detected.

As per DICOM standard, the first 128 bytes of the file are skipped and then the
bytes are read, if they match 0x4449434D then it is proceed as usual by reading
the file as a DICOM.

When the file is not a DICOM because the magic bytes do not match, the other
methods (i.a. PNG, TIFF) are attempted. If the magic bytes are found but the
file is not a DICOM in the end because it is deemed invalid by the specific
reader, then, also the other methods (i.a. PNG, TIFF) are attempted.